Crate linera_alloy_consensus

linera-alloy-consensus

Ethereum consensus interface.

This crate contains constants, types, and functions for implementing Ethereum EL consensus and communication. This includes headers, blocks, transactions, EIP-2718 envelopes, EIP-2930, EIP-4844, and more.

In general a type belongs in this crate if it is committed to in the EL block header. This includes:

transactions

blocks

headers

receipts

EIP-2718 envelopes.

Provenance

Much of this code was ported from reth-primitives as part of ongoing alloy migrations.

Re-exports

pub use transaction::SignableTransaction;

pub use transaction::Transaction;

pub use transaction::TxEip1559;

pub use transaction::TxEip2930;

pub use transaction::TxEip4844;

pub use transaction::TxEip4844Variant;

pub use transaction::TxEip4844WithSidecar;

pub use transaction::TxEnvelope;

pub use transaction::TxLegacy;

pub use transaction::TxType;

pub use transaction::TypedTransaction;

Modules

constants

Ethereum protocol-related constants

transaction

Transaction types.

utils

Utilities for working with EIP-4844 field elements and implementing SidecarCoder.

Structs

Account

Represents an Account in the account trie.

AnyReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Header

Ethereum Block header

Receipt

Receipt containing result of transaction execution.

ReceiptWithBloom

Receipt with calculated bloom filter.

Sealed

A consensus hashable item, with its memoized hash.

SidecarBuilder

Build a BlobTransactionSidecar from an arbitrary amount of data.

Signed

A transaction with a signature and hash seal.

SimpleCoder

Simple coder that only uses the last 31 bytes of each blob. This is the default coder for the SidecarBuilder.

Enums

ReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

Request

Ethereum execution layer requests.

Constants

EMPTY_OMMER_ROOT_HASH

Ommer root of empty list.

EMPTY_ROOT_HASH

Root hash of an empty trie.

Traits

Sealable

Sealeable objects.

SidecarCoder

A strategy for coding and decoding data into sidecars. Coder instances are responsible for encoding and decoding data into and from the sidecar. They are called by the SidecarBuilder during the ingest, take, and (if c_kzg feature enabled) build methods.

TxReceipt

Receipt is the result of a transaction execution.

Type Aliases

Blob

A Blob serialized as 0x-prefixed hex string

Bytes48

A commitment/proof serialized as 0x-prefixed hex string