Module linera_alloy_consensus::constants

Ethereum protocol-related constants

Constants

DEV_GENESIS_HASH

Testnet genesis hash.

EMPTY_OMMER_ROOT_HASH

Ommer root of empty list.

EMPTY_RECEIPTS

Transactions root of empty receipts set.

EMPTY_ROOT_HASH

Root hash of an empty trie.

EMPTY_TRANSACTIONS

Transactions root of empty transactions set.

EMPTY_WITHDRAWALS

Withdrawals root of empty withdrawals set.

ETH_TO_WEI

Multiplier for converting ether to wei.

FINNEY_TO_WEI

Multiplier for converting finney (milliether) to wei.

GOERLI_BASE_GENESIS

Base goerli genesis hash.

GOERLI_GENESIS_HASH

Goerli genesis hash.

GOERLI_OP_GENESIS

Optimism goerli genesis hash.

GWEI_TO_WEI

Multiplier for converting gwei to wei.

HOLESKY_GENESIS_HASH

Holesky genesis hash.

KECCAK_EMPTY

Keccak256 over empty array.

MAINNET_GENESIS_HASH

The Ethereum mainnet genesis hash.

MAXIMUM_EXTRA_DATA_SIZE

Maximum extra data size in a block after genesis

MGAS_TO_GAS

Multiplier for converting mgas to gas.

SELECTOR_LEN

The first four bytes of the call data for a function call specifies the function to be called.

SEPOLIA_GENESIS_HASH

Sepolia genesis hash.