Struct linera_alloy_consensus::Header

pub struct Header {

    pub parent_hash: B256,
    pub ommers_hash: B256,
    pub beneficiary: Address,
    pub state_root: B256,
    pub transactions_root: B256,
    pub receipts_root: B256,
    pub withdrawals_root: Option<B256>,
    pub logs_bloom: Bloom,
    pub difficulty: U256,
    pub number: BlockNumber,
    pub gas_limit: u128,
    pub gas_used: u128,
    pub timestamp: u64,
    pub mix_hash: B256,
    pub nonce: B64,
    pub base_fee_per_gas: Option<u128>,
    pub blob_gas_used: Option<u128>,
    pub excess_blob_gas: Option<u128>,
    pub parent_beacon_block_root: Option<B256>,
    pub requests_root: Option<B256>,
    pub extra_data: Bytes,

}

Ethereum Block header

Fields
parent_hash: B256

The Keccak 256-bit hash of the parent block’s header, in its entirety; formally Hp.

ommers_hash: B256

The Keccak 256-bit hash of the ommers list portion of this block; formally Ho.

beneficiary: Address

The 160-bit address to which all fees collected from the successful mining of this block be transferred; formally Hc.

state_root: B256

The Keccak 256-bit hash of the root node of the state trie, after all transactions are executed and finalisations applied; formally Hr.

transactions_root: B256

The Keccak 256-bit hash of the root node of the trie structure populated with each transaction in the transactions list portion of the block; formally Ht.

receipts_root: B256

The Keccak 256-bit hash of the root node of the trie structure populated with the receipts of each transaction in the transactions list portion of the block; formally He.

withdrawals_root: Option<B256>

The Keccak 256-bit hash of the withdrawals list portion of this block. https://eips.ethereum.org/EIPS/eip-4895

logs_bloom: Bloom

The Bloom filter composed from indexable information (logger address and log topics) contained in each log entry from the receipt of each transaction in the transactions list; formally Hb.

difficulty: U256

A scalar value corresponding to the difficulty level of this block. This can be calculated from the previous block’s difficulty level and the timestamp; formally Hd.

number: BlockNumber

A scalar value equal to the number of ancestor blocks. The genesis block has a number of zero; formally Hi.

gas_limit: u128

A scalar value equal to the current limit of gas expenditure per block; formally Hl.

gas_used: u128

A scalar value equal to the total gas used in transactions in this block; formally Hg.

timestamp: u64

A scalar value equal to the reasonable output of Unix’s time() at this block’s inception; formally Hs.

mix_hash: B256

A 256-bit hash which, combined with the nonce, proves that a sufficient amount of computation has been carried out on this block; formally Hm.

nonce: B64

A 64-bit value which, combined with the mixhash, proves that a sufficient amount of computation has been carried out on this block; formally Hn.

base_fee_per_gas: Option<u128>

A scalar representing EIP1559 base fee which can move up or down each block according to a formula which is a function of gas used in parent block and gas target (block gas limit divided by elasticity multiplier) of parent block. The algorithm results in the base fee per gas increasing when blocks are above the gas target, and decreasing when blocks are below the gas target. The base fee per gas is burned.

blob_gas_used: Option<u128>

The total amount of blob gas consumed by the transactions within the block, added in EIP-4844.

excess_blob_gas: Option<u128>

A running total of blob gas consumed in excess of the target, prior to the block. Blocks with above-target blob gas consumption increase this value, blocks with below-target blob gas consumption decrease it (bounded at 0). This was added in EIP-4844.

parent_beacon_block_root: Option<B256>

The hash of the parent beacon block’s root is included in execution blocks, as proposed by EIP-4788.

This enables trust-minimized access to consensus state, supporting staking pools, bridges, and more.

The beacon roots contract handles root storage, enhancing Ethereum’s functionalities.

requests_root: Option<B256>

The Keccak 256-bit hash of the root node of the trie structure populated with each EIP-7685 request in the block body.

extra_data: Bytes

An arbitrary byte array containing data relevant to this block. This must be 32 bytes or fewer; formally Hx.

Implementations

impl Header

pub fn hash_slow(&self) -> B256

Heavy function that will calculate hash of data and will not save the change to metadata.

Use Header::seal_slow and unlock if you need the hash to be persistent.

pub fn is_empty(&self) -> bool

Checks if the header is empty - has no transactions and no ommers

pub fn ommers_hash_is_empty(&self) -> bool

Check if the ommers hash equals to empty hash list.

pub fn transaction_root_is_empty(&self) -> bool

Check if the transaction root equals to empty root.

pub fn blob_fee(&self) -> Option<u128>

Returns the blob fee for this block according to the EIP-4844 spec.

Returns None if excess_blob_gas is None

pub fn next_block_blob_fee(&self) -> Option<u128>

Returns the blob fee for the next block according to the EIP-4844 spec.

Returns None if excess_blob_gas is None.

See also Self::next_block_excess_blob_gas

pub fn next_block_base_fee( &self, base_fee_params: BaseFeeParams, ) -> Option<u128>

Calculate base fee for next block according to the EIP-1559 spec.

Returns a None if no base fee is set, no EIP-1559 support

pub fn next_block_excess_blob_gas(&self) -> Option<u128>

Calculate excess blob gas for the next block according to the EIP-4844 spec.

Returns a None if no excess blob gas is set, no EIP-4844 support

pub fn size(&self) -> usize

Calculate a heuristic for the in-memory size of the Header.