Struct linera_alloy_consensus::SidecarBuilder

pub struct SidecarBuilder<T = SimpleCoder> { /* private fields */ }

Build a BlobTransactionSidecar from an arbitrary amount of data.

This is useful for creating a sidecar from a large amount of data, which is then split into blobs. It delays KZG commitments and proofs until all data is ready.

Implementations

impl<T> SidecarBuilder<T>
where T: SidecarCoder + Default,

pub fn new() -> SidecarBuilder<T>

Instantiate a new builder and new coder instance.

By default, this allocates space for 2 blobs (256 KiB). If you want to preallocate a specific number of blobs, use SidecarBuilder::with_capacity.

pub fn from_slice(data: &[u8]) -> SidecarBuilder<T>

Create a new builder from a slice of data by calling SidecarBuilder::from_coder_and_data

pub fn with_capacity(capacity: usize) -> SidecarBuilder<T>

Create a new builder with a pre-allocated capacity. This capacity is measured in blobs, each of which is 128 KiB.

impl<T> SidecarBuilder<T>
where T: SidecarCoder,

pub fn from_coder_and_capacity(coder: T, capacity: usize) -> SidecarBuilder<T>

Instantiate a new builder with the provided coder and capacity. This capacity is measured in blobs, each of which is 128 KiB.

pub const fn len(&self) -> usize

Calculate the length of bytes used by field elements in the builder.

This is always strictly greater than the number of bytes that have been ingested.

pub const fn is_empty(&self) -> bool

Check if the builder is empty.

pub fn from_coder_and_data(coder: T, data: &[u8]) -> SidecarBuilder<T>

Create a new builder from a slice of data.

pub fn ingest(&mut self, data: &[u8])

Ingest a slice of data into the builder.

pub fn take( self, ) -> Vec<FixedBytes<linera_alloy_eips::::eip4844::Blob::{constant#0}>>

Take the blobs from the builder, without committing them to a KZG proof.