Trait linera_alloy_consensus::TxReceipt

pub trait TxReceipt<T = Log> {
    // Required methods
    fn status(&self) -> bool;
    fn bloom(&self) -> Bloom;
    fn cumulative_gas_used(&self) -> u128;
    fn logs(&self) -> &[T];

    // Provided method
    fn bloom_cheap(&self) -> Option<Bloom> { ... }
}

Receipt is the result of a transaction execution.

Required Methods

fn status(&self) -> bool

Returns true if the transaction was successful.

fn bloom(&self) -> Bloom

Returns the bloom filter for the logs in the receipt. This operation may be expensive.

fn cumulative_gas_used(&self) -> u128

Returns the cumulative gas used in the block after this transaction was executed.

fn logs(&self) -> &[T]

Returns the logs emitted by this transaction.

Provided Methods

fn bloom_cheap(&self) -> Option<Bloom>

Returns the bloom filter for the logs in the receipt, if it is cheap to compute.