Module linera_alloy_consensus::transaction

Transaction types.

Re-exports

pub use eip4844::TxEip4844;

pub use eip4844::TxEip4844Variant;

pub use eip4844::TxEip4844WithSidecar;

Modules

eip4844

[EIP-4844] constants, helpers, and types.

eip4844_utils

Utilities for working with EIP-4844 field elements and implementing SidecarCoder.

Structs

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

SidecarBuilder

Build a BlobTransactionSidecar from an arbitrary amount of data.

SimpleCoder

Simple coder that only uses the last 31 bytes of each blob. This is the default coder for the SidecarBuilder.

TxEip1559

A transaction with a priority fee (EIP-1559).

TxEip2930

Transaction with an AccessList (EIP-2930).

TxLegacy

Legacy transaction.

Enums

TxEnvelope

The Ethereum EIP-2718 Transaction Envelope.

TxType

Ethereum TransactionType flags as specified in EIPs 2718, 1559, and 2930.

TypedTransaction

The TypedTransaction enum represents all Ethereum transaction request types.

Traits

SidecarCoder

A strategy for coding and decoding data into sidecars. Coder instances are responsible for encoding and decoding data into and from the sidecar. They are called by the SidecarBuilder during the ingest, take, and (if c_kzg feature enabled) build methods.

SignableTransaction

A signable transaction.

Transaction

Represents a minimal EVM transaction.

Type Aliases

Blob

A Blob serialized as 0x-prefixed hex string

Bytes48

A commitment/proof serialized as 0x-prefixed hex string