Struct linera_alloy_consensus::transaction::eip4844::TxEip4844WithSidecar

pub struct TxEip4844WithSidecar {
    pub tx: TxEip4844,
    pub sidecar: BlobTransactionSidecar,
}

EIP-4844 Blob Transaction

A transaction with blob hashes and max blob fee, which also includes the BlobTransactionSidecar. This is the full type sent over the network as a raw transaction. It wraps a TxEip4844 to include the sidecar and the ability to decode it properly.

This is defined in EIP-4844 as an element of a PooledTransactions response, and is also used as the format for sending raw transactions through the network (eth_sendRawTransaction/eth_sendTransaction).

Fields
tx: TxEip4844

The actual transaction.

sidecar: BlobTransactionSidecar

The sidecar.

Implementations

impl TxEip4844WithSidecar

pub const fn from_tx_and_sidecar( tx: TxEip4844, sidecar: BlobTransactionSidecar, ) -> Self

Constructs a new TxEip4844WithSidecar from a TxEip4844 and a BlobTransactionSidecar.

pub const fn tx_type(&self) -> TxType

Get the transaction type.

pub const fn tx(&self) -> &TxEip4844

Get access to the inner tx TxEip4844.

pub const fn sidecar(&self) -> &BlobTransactionSidecar

Get access to the inner sidecar BlobTransactionSidecar.

pub fn into_sidecar(self) -> BlobTransactionSidecar

Consumes the TxEip4844WithSidecar and returns the inner sidecar BlobTransactionSidecar.

pub fn into_parts(self) -> (TxEip4844, BlobTransactionSidecar)

Consumes the TxEip4844WithSidecar and returns the inner TxEip4844 and BlobTransactionSidecar.

pub fn encode_with_signature_fields( &self, signature: &Signature, out: &mut dyn BufMut, )

Encodes the transaction from RLP bytes, including the signature. This does not encode a tx type byte or string header.

This does encode a list header and include a signature.

This encodes the following: rlp([tx_payload, blobs, commitments, proofs])

where tx_payload is the RLP encoding of the TxEip4844 transaction fields: rlp([chain_id, nonce, max_priority_fee_per_gas, ..., v, r, s])