Enum linera_alloy_consensus::transaction::TxEnvelope

#[non_exhaustive]
pub enum TxEnvelope {
    Legacy(Signed<TxLegacy>),
    Eip2930(Signed<TxEip2930>),
    Eip1559(Signed<TxEip1559>),
    Eip4844(Signed<TxEip4844Variant>),
}

The Ethereum EIP-2718 Transaction Envelope.

Note:

This enum distinguishes between tagged and untagged legacy transactions, as the in-protocol merkle tree may commit to EITHER 0-prefixed or raw. Therefore we must ensure that encoding returns the precise byte-array that was decoded, preserving the presence or absence of the TransactionType flag.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Legacy(Signed<TxLegacy>)

An untagged TxLegacy.

Eip2930(Signed<TxEip2930>)

A TxEip2930 tagged with type 1.

Eip1559(Signed<TxEip1559>)

A TxEip1559 tagged with type 2.

Eip4844(Signed<TxEip4844Variant>)

A TxEip4844 tagged with type 3. An EIP-4844 transaction has two network representations: 1 - The transaction itself, which is a regular RLP-encoded transaction and used to retrieve historical transactions..

2 - The transaction with a sidecar, which is the form used to send transactions to the network.

Implementations

impl TxEnvelope

pub fn signature_hash(&self) -> B256

Calculate the signing hash for the transaction.

pub const fn tx_hash(&self) -> &B256

Return the hash of the inner Signed

pub const fn tx_type(&self) -> TxType

Return the TxType of the inner txn.

pub fn inner_length(&self) -> usize

Return the length of the inner txn, without a type byte.