Struct linera_alloy_consensus::transaction::TxEip2930

pub struct TxEip2930 {
    pub chain_id: ChainId,
    pub nonce: u64,
    pub gas_price: u128,
    pub gas_limit: u128,
    pub to: TxKind,
    pub value: U256,
    pub access_list: AccessList,
    pub input: Bytes,
}

Transaction with an AccessList (EIP-2930).

Fields
chain_id: ChainId

Added as EIP-pub 155: Simple replay attack protection

nonce: u64

A scalar value equal to the number of transactions sent by the sender; formally Tn.

gas_price: u128

A scalar value equal to the number of Wei to be paid per unit of gas for all computation costs incurred as a result of the execution of this transaction; formally Tp.

As ethereum circulation is around 120mil eth as of 2022 that is around 120000000000000000000000000 wei we are safe to use u128 as its max number is: 340282366920938463463374607431768211455

gas_limit: u128

A scalar value equal to the maximum amount of gas that should be used in executing this transaction. This is paid up-front, before any computation is done and may not be increased later; formally Tg.

to: TxKind

The 160-bit address of the message call’s recipient or, for a contract creation transaction, ∅, used here to denote the only member of B0 ; formally Tt.

value: U256

A scalar value equal to the number of Wei to be transferred to the message call’s recipient or, in the case of contract creation, as an endowment to the newly created account; formally Tv.

access_list: AccessList

The accessList specifies a list of addresses and storage keys; these addresses and storage keys are added into the accessed_addresses and accessed_storage_keys global sets (introduced in EIP-2929). A gas cost is charged, though at a discount relative to the cost of accessing outside the list.

input: Bytes

Input has two uses depending if transaction is Create or Call (if to field is None or Some). pub init: An unlimited size byte array specifying the EVM-code for the account initialisation procedure CREATE, data: An unlimited size byte array specifying the input data of the message call, formally Td.

Implementations

impl TxEip2930

pub fn size(&self) -> usize

Calculates a heuristic for the in-memory size of the TxEip2930 transaction.

pub const fn tx_type(&self) -> TxType

Get transaction type.